refactored approve task
