+++ packages/core/src/deployments/index.ts
@@ -1,0 +1,11 @@
+import { ethers } from 'ethers'
+
+export const getFinalDeploymentTxnHash = async (
+  ChugSplashManager: ethers.Contract,
+  bundleId: string
+): Promise<string> => {
+  const [finalDeploymentEvent] = await ChugSplashManager.queryFilter(
+    ChugSplashManager.filters.ChugSplashBundleCompleted(bundleId)
+  )
+  return finalDeploymentEvent.transactionHash
+}
