+++ packages/core/src/index.ts
@@ -6,3 +6,5 @@ export * from './fund'
 export * from './constants'
 export * from './tasks'
 export * from './messages'
+export * from './deployments'
+export * from './execution'
